--- src/components/containers/repositories/RepositoryModel.ts
@@ -1,6 +1,0 @@
-export interface RepositoryModel {
-  icon: string
-  name: string
-  url: string
-  description: string
-}
