+++ src/components/shared/social/Social.module.css
@@ -8,16 +8,25 @@
   text-align: center;
 }
 
-.link { display: inherit; }
+.link { 
+  display: inherit; 
+  position: relative;
+}
 
 .link,
-.logo { 
+.social { 
   width: 25px;
   height: 25px;
 }
 
+.social {
+  position: relative;
+  top: 0;
+  transition: var(--transition);
+}
+
+.link:hover .social { top: 1px; }
+
 @media (min-width: breakpoint-md) {
-  .list {
-    max-width: 150px;
-  }
+  .list { max-width: 150px; }
 }
